server/requirements.txt
Django==2.0.7
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
psycopg2==2.7.5
PyJWT==1.6.4
pytz==2018.5
